pyproject.toml
[tool.poetry]
name = "events"
version = "0.1.0"
description = ""
authors = ["Kirill Sumorokov <sumorokov.k@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^21.4.0"
config = { path = "../config/", develop = true }
marshmallow = "^3.17.0"
click = "^8.1.3"
structlog = "^21.5.0"
ujson = "^5.3.0"
aio-pika = "^8.0.3"
aiojobs = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
coverage = "^6.4.1"
flake8 = "^4.0.1"
flake8-aaa = "^0.12.2"
flake8-bugbear = "^22.4.25"
flake8-builtins-unleashed = "^1.3.1"
flake8-comprehensions = "^3.10.0"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-print = "^5.0.0"
flake8-pytest = "^1.3"
ipdb = "^0.13.9"
ipython = "^8.4.0"
mypy = "^0.961"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-black = "^0.3.12"
pytest-cov = "^3.0.0"
pytest-flake8 = "^1.1.1"
pytest-mock = "^3.7.0"
pytest-mypy = "^0.9.1"
pytest-xdist = {extras = ["psutil"], version = "^2.5.0"}
types-setuptools = "^57.4.17"
types-ujson = "^5.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
